refactor: update `useWindowSize` hook file name
